Capping of entries in the EV whitelist generator
This changes allows re-generating the EV whitelist, excluding
expired certificates and ones that were added after the previous
whitelist was generated.

There's also a small bugfix in case the chain submitted consisted
of a single certificate (so the extra_data is empty). That happens
when a root was submitted to the log.