Continue scanning certificates even if an exception is raised
Previously, scanning of a batch of certificates would be aborted if an
unexpected exception was raised and the results would be discarded.
Now, scanning of the certificate that caused the exception will be
aborted, but scanning of the batch will continue and the partial results
will be reported. All exceptions will be logged.